refactor(core): extract daemon-lease reconciliation from JobStore god file
api_jobs/store.rs was a 3119-line god file (audit god_file) dominated by
one massive impl JobStore. Extract the daemon-lease reconciliation and
recovery cluster — daemon_lease_job_diagnostics,
active_daemon_job_recovery_evidence, and the dead / exact-loss / terminal
/ leaseless reconcile family — into a second impl JobStore block in a
reconciliation.rs submodule.

Convert store.rs to a directory module (git mv preserves history). The
submodule uses 'use super::*' to inherit the parent's local helper
types/functions, and JobStore's fields stay pub(super) so the second impl
reaches them. The few reconcile_*_with_* variants exercised by tests are
pub(crate) (their original api_jobs visibility). External contract
(JobStore, JobHandle, JobRunner, RecoveredTerminalJob) unchanged.

Behavior-preserving: store/mod.rs 3119 -> 2106 lines, reconciliation.rs
1046. Every api_jobs test passes in isolation on this branch. (The full
api_jobs suite has pre-existing order-dependent flakiness — the same
tests fail on clean main with a varying count, and each passes when run
alone; verified by stash. Not introduced here.)
